refactor: 공통으로 사용되는 패키지 구조 변경
